+++ test/ploter.py
@@ -91,8 +91,10 @@ def main():
     directory = create_dir('bs',1000000,'Boxing-ram-v4','[372, 64]','[gamma=0.99][prioritized][simple-reward]',0,1)
     plot_single_exp(directory,'b')
     # baseline with not priority repley buffer
-    # directory = create_dir('bs',2000000,'Boxing-ram-v4','[372, 64]','[gamma=0.99][no-prioritized][simple-reward]',0,1)
-    # plot_single_exp(directory,'k')
+    directory = create_dir('bs',2000000,'Boxing-ram-v4','[372, 64]','[gamma=0.99][no-prioritized][simple-reward]',0,1)
+    plot_single_exp(directory,'k')
+    # directory = create_dir('bs',2000000,'Boxing-ram-v4','[372, 64]','[gamma=0.99][new-prioritized][simple-reward]',3,1)
+    # plot_single_exp(directory,'r')
 
     # directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[prioritized][initial][update-all]',3,1)
     # plot_single_exp(directory,'r')
@@ -103,13 +105,15 @@ def main():
     # directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[prioritized][initial][update-all]',0,1)
     # plot_single_exp(directory,'y')
     
-    directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[prioritized][not-initial-clear][update-all]',0,1)
-    plot_single_exp(directory,'y')
-    directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[prioritized][initial-clear][update-all]',0,1)
-    plot_single_exp(directory,'r')
-    directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[global-time][prioritized][not-initial-clear][update-all]',0,1)
-    plot_single_exp(directory,'g')
+    # directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[prioritized][not-initial-clear][update-all]',0,1)
+    # plot_single_exp(directory,'y')
+    # directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[prioritized][initial-clear][update-all]',0,1)
+    # plot_single_exp(directory,'r')
+    # directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[global-time][prioritized][not-initial-clear][update-all]',3,1)
+    # plot_single_exp(directory,'g')
 
+    directory = create_dir('dt',2000000,'Boxing-ram-v4','[372, 64]','[global-time][single-prioritized][not-initial-clear][update-all]',2,1)
+    plot_single_exp(directory,'g')
 
     # game crazyClimber
     # directory = create_dir('bs',3000000,'CrazyClimber-ram-v4','[376, 64]','[gamma=0.99][prioritized]')